sdks/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cortex-memory"
version = "0.1.0"
description = "Graph memory for AI agents"
license = { text = "MIT" }
requires-python = ">=3.9"
readme = "README.md"
dependencies = [
    "grpcio>=1.60",
    "grpcio-tools>=1.60",
    "protobuf>=4.25",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio"]

[tool.hatch.build.targets.wheel]
packages = ["cortex_memory"]
